PathFinding :
Monster's SubMatrix now has an instance of Path finder that can find the shortest path. pathF.shortestPath is an array of char with instructions of the path to take.

If the submatrix is a size of 7 with no wall, the amount of possible path is too big and freezes the game